angular throws DEBUG error when using widget (#2697)
OKTA-516464 fixes esm build
