+++ .bacon.yml
@@ -76,3 +76,11 @@ test_suites:
       script_name: unit-karma
       criteria: MERGE
       queue_name: small
+    - name: verify-registry-install
+      prereq_test_suite_name: publish
+      script_path: /root/okta/okta-signin-widget/scripts
+      sort_order: '10'
+      timeout: '20'
+      script_name: verify-registry-install
+      criteria: MERGE
+      queue_name: small
