+++ rollup.config.js
@@ -1,50 +1,19 @@
-import resolve from '@rollup/plugin-node-resolve';
-import babel from '@rollup/plugin-babel';
-import commonjs from '@rollup/plugin-commonjs';
-import json from '@rollup/plugin-json';
-import { visualizer } from 'rollup-plugin-visualizer';
+import replace from '@rollup/plugin-replace';
+import baseRollupConfig from './rollup.common';
 
-const extensions = ['.js', '.ts'];
-
-// External dependencies should also exist in the package.json dependencies
-// Users of the ESM module will install their own copy
-const external = [
-  '@okta/okta-auth-js',
-  '@sindresorhus/to-milliseconds',
-  'clipboard',
-  'cross-fetch',
-  'parse-ms',
-  'q',
-  'u2f-api-polyfill'
+const plugins = [
+  replace({
+    values: {
+      DEBUG: true
+    },
+    preventAssignment: true
+  }),
+  ...baseRollupConfig.plugins,
 ];
 
-const input = 'src/index.ts';
-const output = {
-  dir: './target/esm',
-  format: 'es',
-  sourcemap: true,
-  preserveModules: true
+const config = {
+  ...baseRollupConfig,
+  plugins
 };
-let plugins = [
-  babel({
-    babelHelpers: 'bundled',
-    extensions
-  }),
-  commonjs(),
-  json(),
-  resolve({
-    extensions,
-    browser: true
-  }),
-  visualizer({
-    filename: './target/esm/okta-sign-in.html',
-    template: 'treemap' // sunburst | treemap | network
-  })
-];
 
-export default {
-  input,
-  output,
-  plugins,
-  external
-};
+export default config;
